+++ packages/domain-post/src/entities/post.entity.ts
@@ -1,6 +1,9 @@
+import type { SagaStatus } from '@volontariapp/shared';
+
 export class PostEntity {
   id!: string;
   authorId!: string;
+  saga_status!: SagaStatus;
   title!: string;
   content!: string;
   createdAt!: Date;