+++ packages/shared/src/enums/index.ts
@@ -2,3 +2,4 @@ export * from './tags-enum.js';
 export * from './user-roles-enum.js';
 export * from './service-type.enum.js';
 export * from './streams.enum.js';
+export * from './saga-status.enum.js';