+++ packages/shared/src/enums/saga-status.enum.ts
@@ -1,0 +1,5 @@
+export enum SagaStatus {
+  PENDING = 'PENDING',
+  DONE = 'DONE',
+  CANCEL = 'CANCEL',
+}