+++ packages/database/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.3.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/shared@0.7.0
+  - @volontariapp/messaging@2.7.7
+
 ## 3.3.7
 
 ### Patch Changes