+++ packages/domain-event/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 3.5.0
+
+### Minor Changes
+
+- saga status added
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/shared@0.7.0
+  - @volontariapp/database@3.3.8
+  - @volontariapp/messaging@2.7.7
+  - @volontariapp/outbox@0.9.26
+
 ## 3.4.6
 
 ### Patch Changes